sfxmp: complete rewrite with asynchronous model
This commit is very large and includes many things, sorry about this.
Long story short, sfxmp now is asynchronous, includes an asynchronous
VideoToolbox implementation derived from the FFmpeg implementation, and
is also licensed to LGPL (because of previous point).

Along with many other small changes that weren't easy to split.

Note: the code is still experimental.