msg: prevent double free on MSG_FRAME
This fixes a regression due to a side effect since 74f899485.

In op_seek(), after the seek request is sent, the message may never
return back. In this case, the seek_msg may contain a previously free'd
MSG_FRAME, which the op_seek() caller (the control thread) may free
again, leading to a crash due the double free if msg->data is not NULL.