decoder_ffmpeg: enable delay_flush for MediaCodec decoders
With the upcoming FFmpeg release (> 3.4), the decoders based on
MediaCodec will change their default behaviour and use a non-blocking
flush which discard all the output buffers retained by the user.

We need to active the new delay_flush option to keep the current
behaviour which delay the flush command until all the retained buffers
have been rendered or discarded.