internals: use stream timebase for improved frame picking accuracy
The timestamps of a video stream at 60 FPS muxed with a 1/1000 time base
(typically a MKV) will look like this:
    0, 17, 33, 50 ...

Internally, sxplayer uses AV_TIME_BASE (FFmpeg timebase, typically used
for the seeking API) which is 1/1000000 (microseconds). Which means the
rescaled timestamps will look like the following:
    0, 17000, 33000, 50000 ...

The second frame with pts=17 means it should be displayed at
t=17*stream_timebase or t=17000*AV_TIMEBASE, 0.017 in both cases.

Meanwhile, a user willing to get video frames at 60Hz will request the
following floats (in double precision):
    0/60, 1/60, 2/60, 3/60 ...

Expressed in AV_TIME_BASE, the values look like this:
    0, 16667, 33333, 50000 ...

This means the comparison between the time and the PTS will not match
in many cases. For the 3rd frame (33000 vs 33333), it's not problematic
because the user time is between frame 3 and 4 so it will display the
correct frame anyway (the 3rd). But in the case of the 2nd frame, the
requested time (16667) is smaller than the rounded PTS (17000), causing
the first frame to be raised again instead.

To solve this issue, this commit makes sure the user requested time is
expressed (generally downscaled) in the stream time base unit when
making timing comparison.

This is a sensible commit since the sxplayer non blocking architecture
needs to deal with timestamps when the media is still not opened (and
thus we do not know the stream time base in many cases). That's the
main reason why the arbitrary AV_TIME_BASE was picked in the first place
and used everywhere.