filtering: fix too strict trim duration check
If the user requested time is past the trim_duration a seek at
t=trim_duration will be made and no frame before this time will be
queued by mod_decoding. This means the following module (mod_filtering)
will not see any frame before the one at t=trim_duration. At this point,
if it decides to drop any frame with t>=trim_duration, no frame will
ever reach the exit. As a result, we need to allow at least
t=trim_duration so the "last" frame of the video will be shown.