decoder_ffmpeg: properly drain the decoder
After receiving a flush packet the remaining buffered frames of the
decoder should be drained using avcodec_receive_frame() until
AVERROR_EOF is returned (indicating that there will be no more output
frames).

To drain the remaining frames, an input buffer with the EOF flag must be
queued to MediaCodec. Input buffers are acquired from MediaCodec but
none can be available (EAGAIN) depending on the current state of the codec.

As a result, it is the responsibility of the final user to handle EAGAIN
in the flush process and wait explicitly for AVERROR_EOF.