decoder: fix multiple audio frames without PTS
This pattern occured in a 3GPP file with an AMR-NB audio stream, with 5
frames decoded per packet, where only the 1st frame had a PTS. In
sxplayer, when the frame PTS is unknown, we drop it (because the timing
information is mandatory by design). But since we have the context of
the packet in a single loop, it's easy to infer the other PTS for this
case.

We can't do that easily at a higher level (meaning interpolating between
packets as well) because the packet order is undefined (typically in
case of a seek).

This simple heuristic is similar to what ffmpeg (the tool) does.