async: do not log an error on lazy image seeking
This error logging was introduced in fc564594 to make it consistent with
the case where a skip option was explicitely specified to a media image.

The log message removed in this commit is the 2nd scenario: it happens
by side effect when doing a sxplayer_get_frame() at t≠0. For API users
(typically node.gl with a TimeRange above a Media node) this is a legit
use case because the time progresses without ever considering the type
of media below. We consider here that it's sxplayer responsiblity to
transparently handles images.