mod_filtering: fix FFT scaling
FFmpeg FFT API states the following on fft_calc (which is used by
rdft_calc): "No 1.0/sqrt(n) normalization is done".

Meaning if we want to normalize the output, we would have to multiply
the real and imaginary parts of each bin (because that's how we scale a
complex). So given s=1/sqrt(n), we have:

    MAG(i) = √((re*s)²+(im*s)²)
           = √(re²*s²+im²*s²)
           = √((re²+im²)*s²)
           = √(re²+im²)*s