ci_win: use pkg-config from Chocolatey
Meson blacklisted the detection of the broken wrapper
Strawberry\perl\bin\pkg-config.BAT starting meson 0.60. This causes the
pkg-config detection to start failing on our Windows CI because it is the only
pkg-config implementation available on the base system. In order to pull a
proper implementation of pkg-config, we use the Chocolatey package manager.