decoder_vt: honor video/full range
We request VT to output an explicit pixel format which we think is the
one that is the less likely to involve a internal pixel convert within
VT.

Here, the range usually comes from the container (or could be probed
from the codec in certain conditions, but not all the time). We expected
it to match what's happening in the bitstream most of the time.

A mismatch may happen for example when the container and the bitstream
disagree, or if the bitstream changes its range mid-stream. In either
case, VT will do a convert to match our requested range, so the output
will still be correct.