api: introduce start_time/end_time trimming interface
Time remapping is usually done with absolute time boundaries because it
avoids the need for arithmetic that can introduce a bias. For that
reason we introduce an absolute end_time parameter to replace the
skip-relative former trim_duration parameter.

Similarly, start_time is introduced as an alias for the skip parameter.

From that point, start_time and end_time absolute controls provide a
consistent interface to trim a playback segment.